+++ test/encode.test.js
@@ -8,11 +8,6 @@ describe('encode', function () {
     const {data, dataRgba, width, height} = await decode(file)
     let buf
 
-    // channels=3
-    buf = encodeSync(data, width, height)
-    buf.should.be.instanceof(Buffer)
-    buf.length.should.above(0)
-
     // channels=4
     buf = encodeSync(dataRgba, width, height)
     buf.should.be.instanceof(Buffer)
@@ -22,11 +17,6 @@ describe('encode', function () {
   it('encode', async () => {
     const {data, dataRgba, width, height} = await decode(file)
     let buf
-
-    // channels=3
-    buf = await encode(data, width, height)
-    buf.should.be.instanceof(Buffer)
-    buf.length.should.above(0)
 
     // channels=4
     buf = await encode(dataRgba, width, height)